+++ app/Http/Requests/Api/Publisher/UpdateRequest.php
@@ -1,0 +1,97 @@
+<?php
+
+namespace App\Http\Requests\Api\Publisher;
+
+use App\Traits\ApiRespons;
+use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Contracts\Validation\Validator;
+use Illuminate\Http\Exceptions\HttpResponseException;
+
+class UpdateRequest extends FormRequest
+{
+    use ApiRespons;
+
+    /**
+     * Indicates if the validator should stop on the first rule failure.
+     *
+     * @var bool
+     */
+    protected $stopOnFirstFailure = false;
+
+    /**
+     * Determine if the user is authorized to make this request.
+     *
+     * @return bool
+     */
+    public function authorize()
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array<string, mixed>
+     */
+    public function rules()
+    {
+        return [
+            'nama' => ['required','string','max:255']
+        ];
+    }
+
+    /**
+     * Custom message for validation
+     *
+     * @return array
+     */
+    public function messages()
+    {
+        return [
+            // 
+        ];
+    }
+
+    /**
+     * Get custom attributes for validator errors.
+     *
+     * @return array
+     */
+    public function attributes()
+    {
+        return [
+            // 
+        ];
+    }
+
+    /**
+     * Prepare the data for validation.
+     *
+     * @return void
+     */
+    protected function prepareForValidation()
+    {
+        $this->merge([
+            // 
+        ]);
+    }
+
+    /**
+     * Custom error message for validation
+     *
+     * @return array
+     */
+    public function failedValidation(Validator $validator)
+    {
+        throw new HttpResponseException(
+            $this->createResponse(500, 'Server Error',
+                [
+                    'error' => $validator->errors()
+                ],
+                [
+                    route('api.author.index')
+                ]
+            )
+        );
+    }
+}
