+++ app/Http/Requests/Api/Siswa/StoreRequest.php
@@ -1,0 +1,105 @@
+<?php
+
+namespace App\Http\Requests\Api\Siswa;
+
+use App\Enums\GenderEnum;
+use App\Traits\ApiRespons;
+use Illuminate\Validation\Rule;
+use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Contracts\Validation\Validator;
+use Illuminate\Http\Exceptions\HttpResponseException;
+
+class StoreRequest extends FormRequest
+{
+    use ApiRespons;
+
+    /**
+     * Indicates if the validator should stop on the first rule failure.
+     *
+     * @var bool
+     */
+    protected $stopOnFirstFailure = false;
+
+    /**
+     * Determine if the user is authorized to make this request.
+     *
+     * @return bool
+     */
+    public function authorize()
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array<string, mixed>
+     */
+    public function rules()
+    {
+        return [
+            'nama' => ['required','string','max:255'],
+            'umur' => ['required','string','max:255'],
+            'kelamin' => ['required','string','max:255',Rule::in(GenderEnum::$gender)],
+            'email' => ['required', 'string','max:255','email:dns','unique:siswas,email'],
+            'nomor_hp' => ['required','string','max:255'],
+            'alamat' => ['required','string'],
+            'kelas' => ['required','string','max:255']
+        ];
+    }
+
+    /**
+     * Custom message for validation
+     *
+     * @return array
+     */
+    public function messages()
+    {
+        return [
+            'kelamin' => 'Data `kelamin` tidak valid, data harus bernilai putra atau putri'
+        ];
+    }
+
+    /**
+     * Get custom attributes for validator errors.
+     *
+     * @return array
+     */
+    public function attributes()
+    {
+        return [
+            // 
+        ];
+    }
+
+    /**
+     * Prepare the data for validation.
+     *
+     * @return void
+     */
+    protected function prepareForValidation()
+    {
+        $this->merge([
+            // 
+        ]);
+    }
+
+    /**
+     * Custom error message for validation
+     *
+     * @return array
+     */
+    public function failedValidation(Validator $validator)
+    {
+        throw new HttpResponseException(
+            $this->createResponse(500, 'Server Error',
+                [
+                    'error' => $validator->errors()
+                ],
+                [
+                    route('api.siswa.index')
+                ]
+            )
+        );
+    }
+}
