+++ app/Http/Requests/Web/Publisher/UpdateRequest.php
@@ -1,18 +1,12 @@
 <?php
 
-namespace App\Http\Requests\Book;
+namespace App\Http\Requests\Web\Publisher;
 
-use App\Traits\ApiRespons;
-use App\Enums\BookStatusEnum;
-use Illuminate\Validation\Rule;
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Contracts\Validation\Validator;
-use Illuminate\Http\Exceptions\HttpResponseException;
 
 class UpdateRequest extends FormRequest
 {
-    use ApiRespons;
-
     /**
      * Indicates if the validator should stop on the first rule failure.
      *
@@ -38,13 +32,7 @@ class UpdateRequest extends FormRequest
     public function rules()
     {
         return [
-            'judul' => ['nullable','string','max:255','unique:books,judul'],
-            'deskripsi' => ['nullable','string'],
-            'author_id' => ['nullable','numeric'],
-            'publisher_id' => ['nullable','numeric'],
-            'category_id' => ['nullable','numeric'],
-            'tanggal_terbit' => ['nullable','date_format:d-m-Y'],
-            'tersedia' => ['nullable','string','max:255',Rule::in(BookStatusEnum::$status)]
+            'nama' => ['required','string','max:255']
         ];
     }
 
@@ -56,7 +44,7 @@ class UpdateRequest extends FormRequest
     public function messages()
     {
         return [
-            'tersedia' => 'Data `tersedia` tidak valid, data harus bernilai Y atau N'
+            // 
         ];
     }
 
@@ -83,23 +71,20 @@ class UpdateRequest extends FormRequest
             // 
         ]);
     }
-
+    
     /**
-     * Custom error message for validation
+     * Custom error response for validation.
      *
      * @return array
      */
     public function failedValidation(Validator $validator)
     {
-        throw new HttpResponseException(
-            $this->createResponse(500, 'Server Error',
-                [
-                    'error' => $validator->errors()
-                ],
-                [
-                    route('api.book.index')
-                ]
-            )
-        );
+        $toast = toastr();
+
+        foreach ($validator->messages()->all() as $message) {
+            $toast->error($message, 'Validation');
+        }
+
+        return $toast;
     }
 }
